[Bronze II] Title: 커트라인, Time: 0 ms, Memory: 13216 KB -BaekjoonHub

+++ 백준/Bronze/25305. 커트라인/커트라인.rs
@@ -1,38 +1,40 @@
 use std::io::{stdin, stdout, BufRead, BufReader, BufWriter, Write};
 
-fn sort(arr: &mut Vec<i64>, n: usize) {
+fn sort(arr: &mut Vec<i32>, n: usize) {
     for i in 1..n {
         let key = arr[i];
-        let mut j = i as isize - 1;
+
+        let mut j = i as i32 - 1;
+
         while j >= 0 && arr[j as usize] < key {
             arr[(j + 1) as usize] = arr[j as usize];
-            j -= 1;
+            j-=1;
         }
         arr[(j + 1) as usize] = key;
     }
 }
-
 fn main() {
     let reader = BufReader::new(stdin().lock());
     let mut writer = BufWriter::new(stdout().lock());
     let mut input = reader.lines();
-    
+
     if let Some(Ok(line)) = input.next() {
         let nk = line
             .trim()
             .split_whitespace()
             .map(|x| x.parse().unwrap())
-            .collect::<Vec<i64>>();
+            .collect::<Vec<usize>>();
         let n = nk[0] as usize;
         let k = nk[1] as usize;
-        
-        if let Some(Ok(line)) = input.next() {
+
+        if let Some(Ok(line)) = input.next(){
             let mut arr = line
                 .trim()
                 .split_whitespace()
                 .map(|x| x.parse().unwrap())
-                .collect::<Vec<i64>>();
+                .collect::<Vec<i32>>();
             sort(&mut arr, n);
+            
             writeln!(writer, "{}", arr[k - 1]).unwrap();
         } else {
             writeln!(writer, "Input not defined").unwrap();
@@ -40,6 +42,6 @@ fn main() {
     } else {
         writeln!(writer, "Input not defined").unwrap();
     }
-    
+
     writer.flush().unwrap();
 }
